Conflicts:
	src/main/java/it/its/testEmployeesDB/FilterCorsConfig.java
	src/main/java/it/its/testEmployeesDB/dao/CitiesDao.java
	src/main/java/it/its/testEmployeesDB/dao/EmployeesDao.java
	src/main/java/it/its/testEmployeesDB/services/CitiesService.java
	src/main/java/it/its/testEmployeesDB/services/CitiesServiceImpl.java
	src/main/java/it/its/testEmployeesDB/services/DipendentiServiceImpl.java
	src/main/java/it/its/testEmployeesDB/services/EmployeesService.java
	src/main/java/it/its/testEmployeesDB/services/EmployeesServiceImpl.java
	src/main/java/it/its/testEmployeesDB/services/ProvincesService.java
	src/main/java/it/its/testEmployeesDB/services/RegionsService.java
	src/main/java/it/its/testEmployeesDB/services/RegionsServiceImpl.java


Merge n 2, prova risoluzione conflitti

+++ src/main/java/it/its/testEmployeesDB/controller/EmployeesController.java
@@ -37,7 +37,7 @@ public class EmployeesController {
 	EmployeesService employeesService;
 	
 	@GetMapping(produces = "application/json", value="/fetchOnce/employee/{idEmployees}")
-	public BaseResponseDto<EmployeesDto> SelOnce(@PathVariable("idEmployees") long idEmployees){
+	public BaseResponseDto<EmployeesDto> SelOnce(@PathVariable("idEmployees") int idEmployees){
 		BaseResponseDto<EmployeesDto> response = new BaseResponseDto<>();
 		
 		Optional<EmployeesDao> dipendenti = dipendentiService.SelOnce(idEmployees);
